requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
torch==2.3.0
transformers==4.41.1
